Refactor commands into separate classes for maintainability
Commands such as "size", "menu", "toggle", and others were moved to dedicated classes, improving modularity and readability. The main "GoPaintCommand" class was streamlined to aggregate these subcommands. This change enhances the structure, making it easier to extend and maintain the code.
