+++ src/main/java/net/thenextlvl/gopaint/command/ToggleCommand.java
@@ -1,0 +1,29 @@
+package net.thenextlvl.gopaint.command;
+
+import com.mojang.brigadier.Command;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import com.mojang.brigadier.context.CommandContext;
+import io.papermc.paper.command.brigadier.CommandSourceStack;
+import io.papermc.paper.command.brigadier.Commands;
+import net.thenextlvl.gopaint.GoPaintPlugin;
+import org.bukkit.entity.Player;
+import org.jspecify.annotations.NullMarked;
+
+@NullMarked
+class ToggleCommand {
+    public static LiteralArgumentBuilder<CommandSourceStack> create(GoPaintPlugin plugin) {
+        return Commands.literal("toggle")
+                .requires(stack -> stack.getSender() instanceof Player)
+                .executes(context -> toggle(context, plugin));
+    }
+
+    private static int toggle(CommandContext<CommandSourceStack> context, GoPaintPlugin plugin) {
+        var player = (Player) context.getSource().getSender();
+        var settings = plugin.brushController().getBrushSettings(player);
+        settings.setEnabled(!settings.isEnabled());
+        var message = settings.isEnabled() ? "command.gopaint.brush.enabled"
+                : "command.gopaint.brush.disabled";
+        plugin.bundle().sendMessage(player, message);
+        return Command.SINGLE_SUCCESS;
+    }
+}
